requirements.txt
Django
opencv-python
easyocr
numpy
Pillow
torch
torchvision
pandas
numpy

